lg_analysis v8 before marge 210821

+++ include/LG/E16ANA_LGBasic.hh
@@ -50,9 +50,9 @@ unordered_map<int, ch_pp*>*  lgdatamap_ip = new unordered_map<int, ch_pp*>;
 ch_pp* GetSpec(uint16_t module, uint16_t block);
 ch_pp* GetSpec(int ip);
 
-void LGWFPeak(float* dat, float* peak, int* peakx, float* timing);
-void LGWFBaseline(float* dat, int peakx, float* baseline, float* baselinerms);
-void LGWFIntegral(float* dat, int peakx, float baseline, float* integral, float* falltime);
+void LGWFPeak(double* dat, double* peak, int* peakx, double* timing);
+void LGWFBaseline(double* dat, int peakx, double* baseline, double* baselinerms);
+void LGWFIntegral(double* dat, int peakx, double baseline, double* integral, int* falltime);
 
   //string file_channelmap();
   //string file_drs4assign();
